+++ hyp3_gamma/insar/ifm_sentinel.py
@@ -292,7 +292,8 @@ def make_parameter_file(mydir, parameter_file_name, alooks, rlooks, dem_source, 
                 s = re.split(r'\s+', t[1])
                 heading = float(s[1])
 
-    orbit_parameters = get_orbit_parameters(reference_file)
+    reference_orbit_parameters = get_orbit_parameters(reference_file)
+    secondary_orbit_parameters = get_orbit_parameters(secondary_file)
 
     reference_file = reference_file.replace(".SAFE", "")
     secondary_file = secondary_file.replace(".SAFE", "")
@@ -300,8 +301,10 @@ def make_parameter_file(mydir, parameter_file_name, alooks, rlooks, dem_source, 
     with open(parameter_file_name, 'w') as f:
         f.write('Reference Granule: %s\n' % reference_file)
         f.write('Secondary Granule: %s\n' % secondary_file)
-        f.write('Pass Direction: %s\n' % orbit_parameters["pass_direction"])
-        f.write('Orbit Number: %s\n' % orbit_parameters["orbitnumber"])
+        f.write('Reference Pass Direction: %s\n' % reference_orbit_parameters["pass_direction"])
+        f.write('Reference Orbit Number: %s\n' % reference_orbit_parameters["orbitnumber"])
+        f.write('Secondary Pass Direction: %s\n' % secondary_orbit_parameters["pass_direction"])
+        f.write('Secondary Orbit Number: %s\n' % secondary_orbit_parameters["orbitnumber"])
         f.write('Baseline: %s\n' % baseline)
         f.write('UTC time: %s\n' % utctime)
         f.write('Heading: %s\n' % heading)
@@ -323,8 +326,6 @@ def make_parameter_file(mydir, parameter_file_name, alooks, rlooks, dem_source, 
         f.write('Phase at reference point: %s\n' % ref_point_info["refoffset"])
         f.write('Azimuth line of the reference point in SAR: %s\n' % coords["row_s"])
         f.write('Range pixel of the reference point in SAR: %s\n' % coords["col_s"])
-        f.write('Row of the reference point in MAP: %s\n' % coords["row_m"])
-        f.write('Column of the reference point in MAP: %s\n' % coords["col_m"])
         f.write('Y of the reference point in MAP: %s\n' % coords["y"])
         f.write('X of the reference point in MAP: %s\n' % coords["x"])
         f.write('Latitude of the reference point: %s\n' % coords["lat"])