+++ CHANGELOG.md
@@ -10,14 +10,13 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [5.1.0](https://github.com/ASFHyP3/hyp3-gamma/compare/v5.0.1...v5.1.0)
 
 ### Added
-* Ascending/descending and orbit number information have been added to the '*.txt' parameter file in InSAR products
+* Ascending/descending and orbit number information have been added to the `*.txt` parameter file in InSAR products
 * The reference point coordinate information (in both SAR and map space) has been
   added to the `.txt` parameter file in InSAR products
 
 ### Changed
 * The highest coherent pixel will be used as the starting point when unwrapping during InSAR
   processing
-
 
 ## [5.0.1](https://github.com/ASFHyP3/hyp3-gamma/compare/v5.0.0...v5.0.1)